move logic to datareposervice
