+++ service/src/main/java/org/databiosphere/workspacedataservice/datarepo/DataRepoConfig.java
@@ -1,6 +1,5 @@
 package org.databiosphere.workspacedataservice.datarepo;
 
-import org.databiosphere.workspacedataservice.dao.RecordDao;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -22,8 +21,8 @@ public class DataRepoConfig {
         }
 
         @Bean
-        public DataRepoDao dataRepoDao(DataRepoClientFactory dataRepoClientFactory, RecordDao recordDao) {
-                return new DataRepoDao(dataRepoClientFactory, recordDao);
+        public DataRepoDao dataRepoDao(DataRepoClientFactory dataRepoClientFactory) {
+                return new DataRepoDao(dataRepoClientFactory);
         }
 
 }
